+++ src/main/java/br/com/fatecmogidascruzes/service/ClienteService.java
@@ -6,7 +6,10 @@ import java.util.List;
 public interface ClienteService {
 
     void adicionarCliente(Cliente cliente);
-    void atualizarCliente(long id, Cliente cliente);
+
+    void atualizarCliente(int id, Cliente cliente);
+
     void excluirCliente(int id);
+
     List<Cliente> buscarCliente(int opcao, String valorBuscar);
 }
