+++ src/main/java/br/com/fatecmogidascruzes/service/FuncionarioService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface FuncionarioService {
 
     void adicionarFuncionario(Funcionario funcionario);
-    void atualizarFuncionario(long id, Funcionario funcionario);
+    void atualizarFuncionario(int id, Funcionario funcionario);
     void excluirFuncionario(int id);
     List<Funcionario> buscarFuncionario(int opcao, String valorBuscar);
 }
